Physics Snapping
Immediately snaps the objects to the player's mouse position, and then when released, lets some velocity continue.

Calculates the magnitude of the vector of mouse movement in the most recent frame, then applies that to the object upon release as a force.

Has a small threshold to reflect inertia.

Easy changes if desired: Increase scale of force applied (based on magnitude of vector), raise or lower inertia threshold, potentially make object snap to object center instead of object pivot.